working on moving frontend.py over to use subprocess_server.py -- working better now.

+++ subprocess_server.py
@@ -257,11 +257,15 @@ class Daemon(object):
             >>> del r
         """
         if pidfile is None:
-            self._pidfile = '%s.pid'%__name__
+            self._pidfile = '%s-%s.pid'%(__name__, port)
         else:
             self._pidfile = pidfile
         if os.path.exists(self._pidfile):
+            max_tries = 10
             while True:
+                max_tries -= 1
+                if max_tries == 0:
+                    break # TODO: here we should just check that it is a zombie
                 try:
                     os.kill(int(open(self._pidfile).read()), signal.SIGKILL)
                     time.sleep(0.05)
